% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_prob.R
\name{const_boundary}
\alias{const_boundary}
\title{Calculate the constant boundary value given a boundary crossing probability.}
\usage{
const_boundary(alpha, d, m_upper = 1000L)
}
\arguments{
\item{alpha}{An upper bound on the boundary crossing probability (positive numeric in \code{[1e-16,0.5]}).}

\item{d}{Bregman divergence between the null and alternative spaces (positive numeric).}

\item{m_upper}{Upper bound on the grid-search for m (default = 1e+3L).}
}
\value{
Constant threshold for GLR-lik statistic which makes the boundary crossing probability upper bounded by \code{alpha} based on Theorem 1.
}
\description{
Constant threshold for well-seperated case.
}
\examples{
const_boundary(0.05, .1)
const_boundary(0.025, 1e-3, m_upper = 100L)
}
