% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_prob.R
\name{const_boundary_lorden}
\alias{const_boundary_lorden}
\title{Constant boundary value given a boundary crossing probability by using Lorden's bound.}
\usage{
const_boundary_lorden(alpha, d)
}
\arguments{
\item{alpha}{An upper bound on the boundary crossing probability (positive numeric in \code{[1e-16,0.5]}).}

\item{d}{Bregman divergence between the null and alternative spaces (positive numeric).}
}
\value{
Constant threshold for GLR-lik statistic which makes the boundary crossing probability upper bounded by \code{alpha} based on the Lorden's bound.
}
\description{
\code{const_boundary_lorden} is used to compute the constant threshold for the sequential GLR test with the well-separated alternative space.
}
\examples{
const_boundary_lorden(0.05, .1)
const_boundary_lorden(0.025, 1e-3)
}
