% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_prob.R
\name{cross_prob}
\alias{cross_prob}
\title{Upper bound on the boundary crossing probability for the sequential GLR-like test.}
\usage{
cross_prob(g, d, nmin = 1L, eta = 1.1)
}
\arguments{
\item{g}{Constant threshold (positive numeric).}

\item{d}{Bregman divergence between the null and alternative spaces (positive numeric).}

\item{nmin}{Minimum sample size of the test (default = 1L).}

\item{eta}{Peeling size (default = 1.1).}
}
\value{
The upper bound on the boundary crossing probability based on Theorem 1 and corresponding K value.
}
\description{
\code{cross_prob} is used to compute the constant threshold for the sequential GLR-like test with the well-separated alternative space case.
}
