% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_seq.R
\name{cross_prob_cs}
\alias{cross_prob_cs}
\title{Upper bound on the boundary crossing probability for confidence sequences.}
\usage{
cross_prob_cs(g, nmax, nmin = 1L, eta = 1.1)
}
\arguments{
\item{g}{Constant threshold (positive numeric).}

\item{nmax}{Upper bound of the target time interval}

\item{nmin}{Lower bound of the target time interval (default = 1L)}

\item{eta}{Peeling size (default = 1.1).}
}
\value{
The upper bound on the boundary crossing probability for confidence sequences and corresponding K value.
}
\description{
\code{cross_prop_cs} is used to compute the constant threshold for the GLR-like confidence sequences with finite target time interval.
}
