% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_prob.R
\name{cross_prob_lorden}
\alias{cross_prob_lorden}
\title{Upper bound on the boundary crossing probability by using Lorden's formula.}
\usage{
cross_prob_lorden(g, d)
}
\arguments{
\item{g}{Constant threshold (positive numeric).}

\item{d}{Bregman divergence between the null and alternative spaces (positive numeric).}
}
\value{
The upper bound on the boundary crossing probability based on Lorden's formula
}
\description{
\code{cross_prob_lorden} is used to compute the constant threshold for the sequential GLR test with the well-separated alternative space based on Lorden's formula.
}
