% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_prob.R
\name{cross_prob_search2}
\alias{cross_prob_search2}
\title{Calculate the upper bound on the boundary crossing probability by using a grid-search}
\usage{
cross_prob_search2(g, d, eta_upper = g/d, eta_grid_size = 0.05)
}
\arguments{
\item{g}{Constant threshold (positive numeric).}

\item{d}{Bregman divergence between the null and alternative spaces (positive numeric).}

\item{eta_upper}{Upper bound on the grid-search for eta (default = g/d).}

\item{eta_grid_size}{Grid size for the grid-search for eta (default = 0.05).}
}
\value{
Grid-search result of the upper bound on the boundary crossing probability based on Theorem 1.
}
\description{
Constant threshold for well-seperated case.
}
\examples{
cross_prob_search2(1, 1)
cross_prob_search2(2, 1, eta_upper = 5, eta_grid_size = 0.5)
}
