% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_prob_no_sep.R
\name{cross_prob_varying}
\alias{cross_prob_varying}
\title{Calculate the upper bound on the boundary crossing probability.}
\usage{
cross_prob_varying(g_fun = NULL, exp_g_fun = NULL, eta = 1.1, k_max = 10000L)
}
\arguments{
\item{g_fun}{positive function on \eqn{[1,\infty)} (R function object, default is \code{NULL}).}

\item{exp_g_fun}{positive function on \eqn{\mathbb{N} \times [1,\infty)} for the exponent term in the upper bound in Theorem 1 (R function object, default is \code{NULL}). Either \code{g_fun} or \code{exp_g_fun} must be non-null. If both are provided, \code{exp_g_fun} will be used.}

\item{eta}{Peeling size (default = 1.1).}

\item{k_max}{Maximum numbers of summations. (positive integer, default = 1e+4L)}
}
\value{
The upper bound on the boundary crossing probability based on Theorem 1.
}
\description{
Time-varying threshold for the non-separation case.
}
