% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_prob_no_sep.R
\name{cross_prob_varying_search}
\alias{cross_prob_varying_search}
\title{Calculate the upper bound on the boundary crossing probability by using a grid-search}
\usage{
cross_prob_varying_search(
  g_fun = NULL,
  exp_g_fun = NULL,
  k_max = 10000L,
  eta_upper = 2,
  eta_grid_size = 0.01
)
}
\arguments{
\item{g_fun}{positive function on \eqn{[1,\infty)} (R function object, default is \code{NULL}).}

\item{exp_g_fun}{positive function on \eqn{\mathbb{N} \times [1,\infty)} for the exponent term in the upper bound in Theorem 1 (R function object, default is \code{NULL}). Either \code{g_fun} or \code{exp_g_fun} must be non-null. If both are provided, \code{exp_g_fun} will be used.}

\item{k_max}{Maximum numbers of summations. (positive integer, default = 1e+4L)}

\item{eta_upper}{Upper bound on the grid-search for eta (default = 2).}

\item{eta_grid_size}{Grid size for the grid-search for eta (default = 0.01).}
}
\value{
Grid-search result of the upper bound on the boundary crossing probability based on Theorem 1 for time-varying boundary.
}
\description{
Time-varying threshold for the non-separation case.
}
\examples{
g_fun <- function(n) 5 + 2 * log(1 + log(n))
exp_g_fun <- function(k, eta) exp(-5 / eta) / (1 + k * log(eta))^(2/eta)
cross_prob_varying_search(g_fun, eta_grid_size = 0.1)
cross_prob_varying_search(g_fun = NULL, exp_g_fun)
}
