% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_fn_class.R
\name{generate_sub_ber_fn}
\alias{generate_sub_ber_fn}
\title{Pre-defined Bregman divergence functions for sub-Bernulli family.}
\usage{
generate_sub_ber_fn(grid_by = 0.1)
}
\arguments{
\item{grid_by}{The size of grid-window of mean space (default = 0.1).}
}
\value{
A list of pre-defined Bregman divergence functions for sub-Bernulli family.
}
\description{
Pre-defined Bregman divergence functions for sub-Bernulli family.
}
