% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_fn_class.R
\name{generate_sub_Gaussian_fn}
\alias{generate_sub_Gaussian_fn}
\title{Pre-defined psi^* and Bregman divergence functions for sub-Gaussian family.}
\usage{
generate_sub_Gaussian_fn(sig = 1, is_add = TRUE)
}
\arguments{
\item{sig}{The sigma parameter of the sub-Gaussian family (default = 1).}

\item{is_add}{If \code{is_add} is \code{TRUE} then return psi^* functions for \code{SGRL_CI_additive}. Otherwise, return Bregman divergence functions for \code{SGLR_CI}.}
}
\value{
A list of pre-defined psi^* and Bregman divergence functions for sub-Gaussian family.
}
\description{
Pre-defined psi^* and Bregman divergence functions for sub-Gaussian family.
}
