% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitching_mixture_method.R
\name{SGLR_CI}
\alias{SGLR_CI}
\title{GLR-like and Discrete Mixture CS for general sub-psi class}
\usage{
SGLR_CI(
  alpha,
  nmax,
  nmin = 1L,
  m_upper = 1000L,
  breg = function(mu_1, mu_0) {     (mu_1 - mu_0)^2/2 },
  breg_pos_inv = function(d, mu_0) {     mu_0 + sqrt(2 * d) },
  breg_neg_inv = function(d, mu_0) {     mu_0 - sqrt(2 * d) },
  breg_derv = function(z, mu_0) {     z - mu_0 },
  mu_lower = NULL,
  mu_upper = NULL,
  grid_by = NULL,
  n_0 = 1L
)
}
\arguments{
\item{alpha}{An upper bound on the boundary crossing probability (positive numeric in \code{[1e-16,0.5]}).}

\item{nmax}{Upper bound of the target time interval.}

\item{nmin}{Lower bound of the target time interval. (default = 1L)}

\item{m_upper}{Upper bound on the grid-search for m (default = 1e+3L).}

\item{breg}{R function of \eqn{D(\mu_1, \mu_0)} which takes \code{mu_1} and \code{mu_0} as inputs (Default: \eqn{(mu_1 - mu_0)^2 / 2}).}

\item{breg_pos_inv}{R function of inverse of the mapping \eqn{z \mapsto D(z, \mu_0):= d} on \eqn{z > \mu_0} which takes \code{d} and \code{mu_0} as inputs (Default: \eqn{\mu_0+\sqrt{2d}}).}

\item{breg_neg_inv}{R function of inverse of the mapping \eqn{z \mapsto D(z, \mu_0):= d} on \eqn{z < \mu_0} which takes \code{d} and \code{mu_0} as inputs (Default: \eqn{\mu_0 - \sqrt{2d}}).}

\item{breg_derv}{R function of \eqn{\nabla_z D(z, \mu_0)} which takes \code{z} and \code{mu_0} as inputs  (Default: \eqn{z - \mu_0}).}

\item{mu_lower}{Lower bound of the mean parameter space (default = NULL).}

\item{mu_upper}{Upper bound of the mean parameter space (default = NULL).}

\item{grid_by}{The size of grid-window of mean space. Default is \code{NULL}.}

\item{n_0}{Lower bound of the sample size on which test statistics and CI will be computed (default = 1L).}
}
\value{
A list of R functions for GLR-like, stitching and discrete mixture bound which takes sample mean \code{x_bar} and sample size \code{n} as the input and return the anytime-valid confidence interval at \code{n}. The list also contains related quantities to compute these bounds. See ADD_CITE for detailed explanations of these quantities.
\describe{
\item{GLR_like_fn}{R function for the GLR-like bound.}
\item{dis_mix_fn}{R function for the discrete mixture bound.}
\item{log_GLR_like_stat_generator}{R function to generate log of GLR-like statistic minus the threshold.}
\item{log_dis_mart_generator}{R function to generate log of GLR-like statistic minus the threshold.}
\item{alpha}{alpha valud used to construct GLR-like and discrete mixture bound functions}
\item{g}{The boundary value for initial \code{nmin} and \code{nmax}.}
\item{eta}{The eta value used to construct underlying martingales}
\item{K}{The number of LR-like martingales used to construct bounds for \eqn{n \geq n_{\min}} part.}
\item{mu_range}{Minimum and maximum of the mean parater space.}
\item{grid_by}{The size of grid-window of mean space.}
}
}
\description{
\code{SGLR_CI} is used to compute R functions which compute bounds of GLR-like, stitching and discrete mixture confidence sequences for general sub-psi class designed to a finite target time interval. For the additive sub-psi class, we recommend to use \code{SGLR_CI_additive} for more efficient computations.
}
\examples{
SGLR_CI(0.025, 1e+5)
SGLR_CI(0.025, 1e+6, 1e+2)
}
