% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitching_mixture_method.R
\name{SGLR_CI_additive}
\alias{SGLR_CI_additive}
\title{GLR-like, Stitching and Discrete Mixture CS for additive sub-psi class}
\usage{
SGLR_CI_additive(
  alpha,
  nmax = NULL,
  nmin = 1L,
  d = NULL,
  m_upper = 1000L,
  psi_star_inv = function(x) {     sqrt(2 * x) },
  psi_star_derv = function(x) {     x },
  n_0 = 1L
)
}
\arguments{
\item{alpha}{An upper bound on the boundary crossing probability (positive numeric in \code{[1e-16,0.5]}).}

\item{nmax}{Upper bound of the target time interval. If both \code{nmax} and \code{d} are provided, \code{nmax} will be ignored. (default = NULL)}

\item{nmin}{Lower bound of the target time interval. (default = 1L)}

\item{d}{Bregman divergence between the null and alternative spaces. Either \code{d} or \code{nmax} must be specified (default = NULL).}

\item{m_upper}{Upper bound on the grid-search for m (default = 1e+3L).}

\item{psi_star_inv}{R function of \eqn{\psi_{+}^{*-1}} (Default: \eqn{\sqrt{2x}}) .}

\item{psi_star_derv}{R function of \eqn{\nabla\psi^*} (Default: \eqn{x}).}

\item{n_0}{Lower bound of the sample size on which test statistics and CI will be computed (default = 1L).}
}
\value{
A list of R functions for GLR-like, stitching and discrete mixture bound which takes the sample size \code{n} as the input and return the distance from the sample mean to the upper bound of confidence interval at \code{n}. The list also contains related quantities to compute these bounds. See ADD_CITE for detailed explanations of these quantities.
\describe{
\item{GLR_like_fn}{R function for the GLR-like bound.}
\item{stitch_fn}{R function for the stitching bound.}
\item{dis_mix_fn}{R function for the discrete mixture bound.}
\item{log_mtg_fn}{R function for the log of the underlying log super-martingale.}
\item{g}{The constant bouandry value.}
\item{eta}{The eta value used to construct underlying martingales}
\item{K}{The number of LR-like martingales used to construct bounds for \eqn{n \geq n_{\min}} part.}
}
}
\description{
\code{SGLR_CI_additive} is used to compute R functions which compute bounds of GLR-like, stitching and discrete mixture confidence sequences for additive sub-psi class designed to a finite target time interval.
}
\examples{
SGLR_CI_additive(0.025, 1e+5)
SGLR_CI_additive(0.025, 1e+6, 1e+2)
}
