% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_fn.R
\name{test_simul}
\alias{test_simul}
\title{\code{test_simul} is used to conduct a simulation which compare the sample size and detection rates of sequential GLR-lik, discrete mixture test and a fixed sample size test based on the sample mean.}
\usage{
test_simul(
  mu_target,
  mu_true = mu_target,
  mu_0,
  alpha = 0.1,
  beta = 0.1,
  B = 100,
  print_progress = TRUE,
  print_result = FALSE,
  sample_generator = G_sample,
  power_fn = G_pwr,
  fixed_test_fn = z_test,
  seq_test_fn = seq_G_test_generator
)
}
\arguments{
\item{mu_target}{The boundary of the target alternative hypothesis space.}

\item{mu_true}{The underlying mean of the data-generating distribution (default = \code{mu_target}).}

\item{mu_0}{The boundary of the null hypothesis space.}

\item{alpha}{The upper bound of the type-I error (default = \code{0.1})}

\item{beta}{The upper bound of the type-II error (default = \code{0.1}).}

\item{B}{The number of the repeats of the simulation (default = \code{100}).}

\item{print_progress}{A logical value. If \code{print_progress} = \code{TRUE}, a progress bar of the simulation will be printed (default = \code{TRUE}).}

\item{print_result}{A logical value. If \code{print_result} = \code{TRUE}, A short summary of the simulation result will be presented (default = \code{FALSE}).}

\item{sample_generator}{R function to generate i.i.d. random samples. The function must take the number of samples \code{n} and the underlying mean \code{mu_true} as inputs.}

\item{power_fn}{R function to compute the power of the test given the sample size. The function must take the number of samples \code{n}, the boundary of the alternative space \code{mu_target}, the boundary of the null space \code{mu_0}, and type-I error bound \code{alpha}.}

\item{fixed_test_fn}{R function to conduct a fixed sample size test. The function must take the sample mean \code{x_bar}, the number of sample \code{n}, The boundary of the null space \code{mu_0}, and the type-I error bound \code{alpha}. The function must return a positive numeric value if the test reject the null.}

\item{seq_test_fn}{R function to conduct sequential GLR-like and discrete mixture tests. The function must take the type-I error bound \code{alpha}, the upper bound \code{nmax} and the lower bound \code{nmin} of the target interval.}
}
\value{
A list of simulation results and underlying settings.
\describe{
\item{reject_rate}{Estimated probabilities of rejecting the null hypothesis.}
\item{sample_size}{Estimated average sample sizes of testing procedures.}
\item{early_stop_ratio}{Estimated probabilities of tests being stopped earlier than the fixed sample size test.}
\item{reject_list}{A list of vectors representing whether each test reject the null (\code{1}) or not (\code{0}) for each simulation run.}
\item{sample_size_list}{A list of numbers of samples used in each test.}
\item{mu_target}{The boundary of the target alternative hypothesis space.}
\item{mu_true}{The underlying mean of the data-generating distribution.}
\item{mu_0}{The boundary of the null hypothesis space.}
\item{alpha}{The upper bound of the type-I error.}
\item{beta}{The upper bound of the type-II error.}
}
}
\description{
\code{test_simul} is used to conduct a simulation which compare the sample size and detection rates of sequential GLR-lik, discrete mixture test and a fixed sample size test based on the sample mean.
}
\examples{
# For the example of the simulation, please check https://github.com/shinjaehyeok/SGLRT_paper.
}
