% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_seq.R
\name{const_boundary_cs}
\alias{const_boundary_cs}
\title{Constant boundary value given a boundary crossing probability for confidence sequences.}
\usage{
const_boundary_cs(alpha, nmax = NULL, nmin = 1L, d = NULL, m_upper = 1000L)
}
\arguments{
\item{alpha}{An upper bound on the boundary crossing probability (positive numeric in \code{[1e-16,0.5]}).}

\item{nmax}{Upper bound of the target time interval. If both \code{nmax} and \code{d} are provided, \code{nmax} will be ignored. (default = NULL)}

\item{nmin}{Lower bound of the target time interval. (default = 1L)}

\item{d}{Bregman divergence between the null and alternative spaces. Either \code{d} or \code{nmax} must be specified (default = NULL).}

\item{m_upper}{Upper bound on the grid-search for m (default = 1e+3L).}
}
\value{
Constant threshold for GLR-like confidence sequence with level \code{alpha} based on Theorem 7.
}
\description{
\code{const_boundary_cs} is used to compute the constant threshold for the GLR-like confidence sequences with finite target time interval.
}
\examples{
const_boundary_cs(0.05, 1e+6)
const_boundary_cs(0.025, 1e+8, 1e+2, m_upper = 100L)
}
