% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_prob.R
\name{cross_prob}
\alias{cross_prob}
\title{Calculate the upper bound on the boundary crossing probability.}
\usage{
cross_prob(g, d, eta = 1.1)
}
\arguments{
\item{g}{Constant threshold (positive numeric).}

\item{d}{Bregman divergence between the null and alternative spaces (positive numeric).}

\item{eta}{Peeling size (default = 1.1).}
}
\value{
The upper bound on the boundary crossing probability based on Theorem 1 and corresponding K value.
}
\description{
Constant threshold for well-seperated case.
}
