% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_seq.R
\name{cross_prob_cs}
\alias{cross_prob_cs}
\title{Calculate the upper bound on the boundary crossing probability for confidence sequences.}
\usage{
cross_prob_cs(g, nmax, nmin = 1L, eta = 1.1)
}
\arguments{
\item{g}{Constant threshold (positive numeric).}

\item{nmax}{Upper bound of the target time interval}

\item{nmin}{Lower bound of the target time interval (default = 1L)}

\item{eta}{Peeling size (default = 1.1).}
}
\value{
The upper bound on the boundary crossing probability for confidence sequeneces and corresponding K value.
}
\description{
Constant threshold with finite target time interval.
}
\examples{
cross_prob_cs(1, 1e+6)
cross_prob_cs(2, 1e+8, 1e+2, eta = 2)
}
