% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_prob.R
\name{cross_prob_lorden}
\alias{cross_prob_lorden}
\title{Calculate the upper bound on the boundary crossing probability by using      Lorden's formula.}
\usage{
cross_prob_lorden(g, d)
}
\arguments{
\item{g}{Constant threshold (positive numeric).}

\item{d}{Bregman divergence between the null and alternative spaces (positive numeric).}
}
\value{
The upper bound on the boundary crossing probability based on Lorden's formula
}
\description{
Constant threshold for well-seperated case.
}
