% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_prob.R
\name{cross_prob_search}
\alias{cross_prob_search}
\title{Grid-searching the upper bound on the boundary crossing probability for the sequential GLR-like test.}
\usage{
cross_prob_search(g, d, nmin = 1L, m_upper = 1000L)
}
\arguments{
\item{g}{Constant threshold (positive numeric).}

\item{d}{Bregman divergence between the null and alternative spaces (positive numeric).}

\item{nmin}{Minimum sample size of the test (default = 1L).}

\item{m_upper}{Upper bound on the grid-search for m (default = 1e+3L).}
}
\value{
Grid-search result of the upper bound on the boundary crossing probability based on Theorem 1.
}
\description{
\code{cross_prob_search} is used to grid-search the constant threshold for the sequential GLR-like test with the well-separated alternative space case.
}
