% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitching_mixture_method.R
\name{SGLR_CI}
\alias{SGLR_CI}
\title{GLR-like, Stitching and Discrete Mixture CS for additive sub-psi class}
\usage{
SGLR_CI(
  alpha,
  nmax,
  nmin = 1L,
  m_upper = 1000L,
  psi_star = function(x) {     x^2/2 },
  psi_star_inv = function(x) {     sqrt(2 * x) },
  psi_star_derv = function(x) {     x }
)
}
\arguments{
\item{alpha}{An upper bound on the boundary crossing probability (positive numeric in \code{[1e-16,0.5]}).}

\item{nmax}{Upper bound of the target time interval.}

\item{nmin}{Lower bound of the target time interval. (default = 1L)}

\item{m_upper}{Upper bound on the grid-search for m (default = 1e+3L).}

\item{psi_star}{R function of \eqn{\psi^*} (Default: \eqn{x^2 /2} (sub-Gussian with \eqn{\sigma = 1}.))}

\item{psi_star_inv}{R function of \eqn{\psi^{*-1}} (Default: \eqn{\sqrt{2x}}) .}

\item{psi_star_derv}{R function of \eqn{\nabla\psi^*} (Default: \eqn{x}).}
}
\value{
A list of R functions for stitching and discrete mixture bound which takes the sample size \code{n} as the input and return the distance from the sample mean to the upper bound of confidence interval at \code{n}. The list also contains related quantities to compute these bounds. See ADD_CITE for detailed explanations of these quantities.
\describe{
\item{GLR_like_fn}{R function for the GLR-like bound.}
\item{stitch_fn}{R function for the stitching bound.}
\item{dis_mix_fn}{R function for the discrete mixture bound.}
\item{g}{The constant bouandry value.}
\item{eta}{The eta value used to construct underlying martingales}
\item{K}{The number of LR-like martingales used to construct bounds for \eqn{n \geq n_{\min}} part.}
}
}
\description{
\code{SGLR_CI} is used to compute R functions which compute bounds of GLR-like, stitching and discrete mixture confidence sequences desgined to a finite target time interval.
}
\examples{
SGLR_CI(0.025, 1e+5)
SGLR_CI(0.025, 1e+6, 1e+2)
}
